chore(mcp): move ukg ready to new architecture
